# app/Http/Requests/Admin/Post/SearchPostRequest.php
<?php

namespace App\Http\Requests\Admin\Post;

use Illuminate\Foundation\Http\FormRequest;
use Illuminate\Support\Facades\Auth;
use OpenApi\Attributes as OA;
#[OA\Schema(
    schema: "SearchPostRequest",
    required: ['search'],
    properties: [
        new OA\Property(property: "search", type: "string")
    ]
)]
class SearchPostRequest extends FormRequest
{
    /**
     * Determine if the user is authorized to make this request.
     */
    public function authorize(): bool
    {
        return Auth::check();
    }

    /**
     * Get the validation rules that apply to the request.
     *
     * @return array<string, \Illuminate\Contracts\Validation\ValidationRule|array<mixed>|string>
     */
    public function rules(): array
    {
        return [
            'search' => ['string','required','min:3','max:100'],
        ];
    }
}